refactor(common): extract calendar popper placement into a util
getCalendarPopperPlacement is pure geometry with no React dependency, so it
does not belong inside the hook. Moving it to common/utils follows the repo's
util convention and makes the three placement branches directly assertable
without renderHook. The CalendarPopperPlacement type moves to CommonTypes,
matching the convention that utils do not export types inline.
